# book-network-ui/src/app/services/models/api-response-authentication-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AuthenticationResponse } from '../models/authentication-response';
export interface ApiResponseAuthenticationResponse {
  data?: AuthenticationResponse;
  errors?: Array<string>;
  message?: string;
  success?: boolean;
  timestamp?: string;
}