# book-network-ui/src/app/services/models/api-response-string.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApiResponseString {
  data?: string;
  errors?: Array<string>;
  message?: string;
  success?: boolean;
  timestamp?: string;
}