# book-network-ui/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponse {
  archived?: boolean;
  authorName?: string;
  cover?: string;
  id?: number;
  isbn?: string;
  message?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}